feat: added lazy.nvim & lua lsp config
The config will use lazy.nvim as package manager, the lock file will be
static to ensure reproducibility.

Also Lua lsp was configured, with strict static type checking ect, but
is tailored for use on nvim configs, not really for any project. That's
my only use case for lua atleast for now